systemd_depend_*: don't restart services
First, I don't think it's even necessary. If the service wasn't running,
then we don't need to restart it. If it was running, it'll continue
running fine with a new dependency; it'll be picked up on a future
restart.

Second, all of that doesn't even work due to (another) weird way Ansible
treats handlers:

    https://github.com/ansible/ansible/issues/81722